refactor(server): split mcp-proxy.ts below the 1000-line cap (#409)
Extract SseFrameParser into sse-frame-parser.ts and probeDaemon/waitForDaemon
into proxy-daemon-probe.ts. Pure moves with re-exports — all 168 MCP tests pass
unchanged. mcp-proxy.ts drops from 994 to 898 lines.
